test(frontend): icrc-wallet.worker
